Renamed characteristics file, Product breakdown bars now display correct text
