+++ client/src/components/R&R/characteristicsText.js
